Video_models/Density_models/FMDC/requirements.txt
torchsummary==1.5.1
tqdm==4.48.2
scipy==1.6.2
matplotlib==3.5.1
tensorboardX
tensorboard
easydict
pandas
numpy
opencv-python
pyyaml
Pillow==9.5.0
numpy==1.22.0